+++ packages/terra-date-picker/tests/wdio/date-picker-spec.js
@@ -480,9 +480,6 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
   describe('Initial Focus', () => {
     before(() => {
       browser.url('/#/raw/tests/cerner-terra-framework-docs/date-picker/date-picker-default');
-      $('input[name="terra-date-month-date-input"]').setValue('12');
-      $('input[name="terra-date-day-date-input"]').setValue('18');
-      $('input[name="terra-date-year-date-input"]').setValue('1994');
       Terra.hideInputCaret('input[name="terra-date-month-date-input"]');
       Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
       Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
@@ -527,67 +524,74 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
 
   describe('Arrow, Delete and Backspace Navigation', () => {
     describe('Up Arrow', () => {
-      before(() => {
+      beforeEach(() => {
         browser.url('/#/raw/tests/cerner-terra-framework-docs/date-picker/date-picker-default');
-        Terra.hideInputCaret('input[name="terra-date-month-date-input"]');
-        Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
-        Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
       });
 
-      it('Increment month input', () => {
+      it('sets month to next month on month input', () => {
+        Terra.hideInputCaret('input[name="terra-date-month-date-input"]');
         $('input[name="terra-date-month-date-input"]').click();
         browser.keys('ArrowUp');
 
-        Terra.validates.element('up arrow increments month input');
+        const tomorrow = new Date();
+        tomorrow.setDate(tomorrow.getMonth() + 1);
+        expect($('input')).toHaveValue(tomorrow.toISOString().split('T')[0]);
       });
 
-      it('Increment day input', () => {
+      it('sets Date to tomorrow if on day input', () => {
+        Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
         $('input[name="terra-date-day-date-input"]').click();
         browser.keys('ArrowUp');
 
-        Terra.validates.element('up arrow increments day input');
+        const tomorrow = new Date();
+        tomorrow.setDate(tomorrow.getDate() + 1);
+        expect($('input')).toHaveValue(tomorrow.toISOString().split('T')[0]);
       });
 
-      it('Increment year input', () => {
+      it('sets year to last year if on year input', () => {
+        Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
         $('input[name="terra-date-year-date-input"]').click();
         browser.keys('ArrowUp');
 
-        Terra.validates.element('up arrow increments year input');
+        const tomorrow = new Date();
+        tomorrow.setDate(tomorrow.getFullYear() + 1);
+        expect($('input')).toHaveValue(tomorrow.toISOString().split('T')[0]);
       });
     });
 
     describe('Down Arrow', () => {
-      before(() => {
+      beforeEach(() => {
         browser.url('/#/raw/tests/cerner-terra-framework-docs/date-picker/date-picker-default');
+      });
+
+      it('sets month to last month on month input', () => {
         Terra.hideInputCaret('input[name="terra-date-month-date-input"]');
-        Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
-        Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
-      });
-
-      it('Decrement month input', () => {
         $('input[name="terra-date-month-date-input"]').click();
-        browser.keys('10');
-        browser.keys('ArrowLeft');
         browser.keys('ArrowDown');
 
-        Terra.validates.element('down arrow decrements month input');
+        const yesterday = new Date();
+        yesterday.setDate(yesterday.getMonth() - 1);
+        expect($('input')).toHaveValue(yesterday.toISOString().split('T')[0]);
       });
 
-      it('Decrement day input', () => {
+      it('sets Date to yesterday if on day input', () => {
+        Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
         $('input[name="terra-date-day-date-input"]').click();
-        browser.keys('10');
-        browser.keys('ArrowLeft');
         browser.keys('ArrowDown');
 
-        Terra.validates.element('down arrow decrements day input');
+        const yesterday = new Date();
+        yesterday.setDate(yesterday.getDate() - 1);
+        expect($('input')).toHaveValue(yesterday.toISOString().split('T')[0]);
       });
 
-      it('Decrement year input', () => {
+      it('sets year to last year if on year input', () => {
+        Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
         $('input[name="terra-date-year-date-input"]').click();
-        browser.keys('2009');
         browser.keys('ArrowDown');
 
-        Terra.validates.element('down arrow decrements year input');
+        const yesterday = new Date();
+        yesterday.setDate(yesterday.getFullYear() - 1);
+        expect($('input')).toHaveValue(yesterday.toISOString().split('T')[0]);
       });
     });
 
@@ -787,13 +791,13 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           browser.url('/#/raw/tests/cerner-terra-framework-docs/date-picker/date-picker-default');
         });
 
-        it('sets Date to yesterday on month input', () => {
+        it('sets month to last month on month input', () => {
           Terra.hideInputCaret('input[name="terra-date-month-date-input"]');
           $('input[name="terra-date-month-date-input"]').click();
           browser.keys('-');
 
           const yesterday = new Date();
-          yesterday.setDate(yesterday.getDate() - 1);
+          yesterday.setDate(yesterday.getMonth() - 1);
           expect($('input')).toHaveValue(yesterday.toISOString().split('T')[0]);
         });
 
@@ -807,13 +811,13 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           expect($('input')).toHaveValue(yesterday.toISOString().split('T')[0]);
         });
 
-        it('sets Date to yesterday if on year input', () => {
+        it('sets year to last year if on year input', () => {
           Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
           $('input[name="terra-date-year-date-input"]').click();
           browser.keys('-');
 
           const yesterday = new Date();
-          yesterday.setDate(yesterday.getDate() - 1);
+          yesterday.setDate(yesterday.getFullYear() - 1);
           expect($('input')).toHaveValue(yesterday.toISOString().split('T')[0]);
         });
       });
@@ -823,7 +827,7 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           browser.url('/#/raw/tests/cerner-terra-framework-docs/date-picker/date-picker-default');
         });
 
-        it('sets Date to yesterday on month input', () => {
+        it('sets month to last month on month input', () => {
           Terra.hideInputCaret('input[name="terra-date-month-date-input"]');
           $('input[name="terra-date-month-date-input"]').click();
           browser.keys('06');
@@ -831,7 +835,7 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           browser.keys('-');
 
           const yesterday = new Date();
-          yesterday.setDate(yesterday.getDate() - 1);
+          yesterday.setDate(yesterday.getMonth() - 1);
           expect($('input')).toHaveValue(yesterday.toISOString().split('T')[0]);
         });
 
@@ -847,7 +851,7 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           expect($('input')).toHaveValue(yesterday.toISOString().split('T')[0]);
         });
 
-        it('sets Date to yesterday if on year input', () => {
+        it('sets year to last year if on year input', () => {
           Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
           $('input[name="terra-date-year-date-input"]').click();
           browser.keys('2005');
@@ -855,7 +859,7 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           browser.keys('-');
 
           const yesterday = new Date();
-          yesterday.setDate(yesterday.getDate() - 1);
+          yesterday.setDate(yesterday.getFullYear() - 1);
           expect($('input')).toHaveValue(yesterday.toISOString().split('T')[0]);
         });
       });
@@ -865,12 +869,12 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           browser.url('/#/raw/tests/cerner-terra-framework-docs/date-picker/date-picker-start-date');
         });
 
-        it('sets Date to valid date minus 1 day if pressed in month', () => {
+        it('sets Date to valid date minus 1 month if pressed in month', () => {
           Terra.hideInputCaret('input[name="terra-date-month-date-input"]');
           $('input[name="terra-date-month-date-input"]').click();
           browser.keys('-');
 
-          expect($('input')).toHaveValue('2017-03-31');
+          expect($('input')).toHaveValue('2017-03-01');
         });
 
         it('sets Date to valid date minus 1 day if pressed in day', () => {
@@ -881,12 +885,12 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           expect($('input')).toHaveValue('2017-03-31');
         });
 
-        it('sets Date to valid date minus 1 day if pressed in year', () => {
+        it('sets Date to valid date minus 1 year if pressed in year', () => {
           Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
           $('input[name="terra-date-year-date-input"]').click();
           browser.keys('-');
 
-          expect($('input')).toHaveValue('2017-03-31');
+          expect($('input')).toHaveValue('2016-04-01');
         });
       });
 
@@ -910,13 +914,13 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           browser.url('/#/raw/tests/cerner-terra-framework-docs/date-picker/date-picker-default');
         });
 
-        it('sets Date to tomorrow on month input', () => {
+        it('sets month to next month on month input', () => {
           Terra.hideInputCaret('input[name="terra-date-month-date-input"]');
           $('input[name="terra-date-month-date-input"]').click();
           browser.keys('=');
 
           const tomorrow = new Date();
-          tomorrow.setDate(tomorrow.getDate() + 1);
+          tomorrow.setDate(tomorrow.getMonth() + 1);
           expect($('input')).toHaveValue(tomorrow.toISOString().split('T')[0]);
         });
 
@@ -930,13 +934,13 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           expect($('input')).toHaveValue(tomorrow.toISOString().split('T')[0]);
         });
 
-        it('sets Date to tomorrow if on year input', () => {
+        it('sets year to last year if on year input', () => {
           Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
           $('input[name="terra-date-year-date-input"]').click();
           browser.keys('=');
 
           const tomorrow = new Date();
-          tomorrow.setDate(tomorrow.getDate() + 1);
+          tomorrow.setDate(tomorrow.getFullYear() + 1);
           expect($('input')).toHaveValue(tomorrow.toISOString().split('T')[0]);
         });
       });
@@ -946,7 +950,7 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           browser.url('/#/raw/tests/cerner-terra-framework-docs/date-picker/date-picker-default');
         });
 
-        it('sets Date to tomorrow on month input', () => {
+        it('sets month to next month on month input', () => {
           Terra.hideInputCaret('input[name="terra-date-month-date-input"]');
           $('input[name="terra-date-month-date-input"]').click();
           browser.keys('06');
@@ -954,7 +958,7 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           browser.keys('=');
 
           const tomorrow = new Date();
-          tomorrow.setDate(tomorrow.getDate() + 1);
+          tomorrow.setDate(tomorrow.getMonth() + 1);
           expect($('input')).toHaveValue(tomorrow.toISOString().split('T')[0]);
         });
 
@@ -970,7 +974,7 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           expect($('input')).toHaveValue(tomorrow.toISOString().split('T')[0]);
         });
 
-        it('sets Date to tomorrow if on year input', () => {
+        it('sets yesr to last year if on year input', () => {
           Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
           $('input[name="terra-date-year-date-input"]').click();
           browser.keys('2005');
@@ -978,7 +982,7 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           browser.keys('=');
 
           const tomorrow = new Date();
-          tomorrow.setDate(tomorrow.getDate() + 1);
+          tomorrow.setDate(tomorrow.getFullYear() + 1);
           expect($('input')).toHaveValue(tomorrow.toISOString().split('T')[0]);
         });
       });
@@ -988,12 +992,12 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           browser.url('/#/raw/tests/cerner-terra-framework-docs/date-picker/date-picker-start-date');
         });
 
-        it('sets Date to valid date plus 1 day if pressed in month', () => {
+        it('sets Date to valid date plus 1 month if pressed in month', () => {
           Terra.hideInputCaret('input[name="terra-date-month-date-input"]');
           $('input[name="terra-date-month-date-input"]').click();
           browser.keys('=');
 
-          expect($('input')).toHaveValue('2017-04-02');
+          expect($('input')).toHaveValue('2017-05-01');
         });
 
         it('sets Date to valid date plus 1 day if pressed in day', () => {
@@ -1004,12 +1008,12 @@ Terra.describeViewports('Date Picker', ['medium'], () => {
           expect($('input')).toHaveValue('2017-04-02');
         });
 
-        it('sets Date to valid date plus 1 day if pressed in year', () => {
+        it('sets Date to valid date plus 1 year if pressed in year', () => {
           Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
           $('input[name="terra-date-year-date-input"]').click();
           browser.keys('=');
 
-          expect($('input')).toHaveValue('2017-04-02');
+          expect($('input')).toHaveValue('2018-04-01');
         });
       });